Add proper authentication redirects
