+++ package.json
@@ -69,6 +69,7 @@
     "react-window": "^1.8.6",
     "resize-observer-polyfill": "^1.5.1",
     "serialize-javascript": "^6.0.0",
+    "server-only": "^0.0.1",
     "universal-cookie": "^4.0.3",
     "uuid": "^8.3.0",
     "xstate": "^4.23.1",
